extract types to avoid circular dependencies
