+++ plugins/graphql/src/app/resolver.ts
@@ -1,10 +1,4 @@
-import { CatalogApi } from "@backstage/catalog-client";
-import { Loader } from "./loaders";
-
-export interface ResolverContext {
-  loader: Loader
-  catalog: CatalogApi
-}
+import type { ResolverContext } from "./types";
 
 export async function resolveType({ id }: { id: string }, { loader }: ResolverContext): Promise<string | null> {
   const entity = await loader.load(id);
